// src/interface/event.interface.ts
// Generated by https://quicktype.io

export interface IEvent {
    groupedbycourse: Groupedbycourse[];
}

export interface Groupedbycourse {
    events: Event[];
    firstid: number | null;
    lastid: number | null;
    courseid: number;
}

export interface Event {
    id: number;
    name: string;
    description: string;
    descriptionformat: number;
    location: string;
    categoryid: null;
    groupid: null;
    userid: number | null;
    repeatid: null;
    eventcount: null;
    component: string;
    modulename: string;
    activityname: string;
    activitystr: string;
    instance: number;
    eventtype: string;
    timestart: number;
    timeduration: number;
    timesort: number;
    timeusermidnight: number;
    visible: number;
    timemodified: number;
    overdue: boolean;
    icon: Icon;
    course: Course;
    subscription: Subscription;
    canedit: boolean;
    candelete: boolean;
    deleteurl: string;
    editurl: string;
    viewurl: string;
    formattedtime: string;
    isactionevent: boolean;
    iscourseevent: boolean;
    iscategoryevent: boolean;
    groupname: null;
    normalisedeventtype: string;
    normalisedeventtypetext: string;
    action: Action;
    purpose: string;
    url: string;
}

export interface Action {
    name: string;
    url: string;
    itemcount: number;
    actionable: boolean;
    showitemcount: boolean;
}

export interface Course {
    id: number;
    fullname: string;
    shortname: string;
    idnumber: string;
    summary: string;
    summaryformat: number;
    startdate: number;
    enddate: number;
    visible: boolean;
    showactivitydates: boolean;
    showcompletionconditions: null;
    fullnamedisplay: string;
    viewurl: string;
    courseimage: string;
    progress: number;
    hasprogress: boolean;
    isfavourite: boolean;
    hidden: boolean;
    showshortname: boolean;
    coursecategory: string;
}

export interface Icon {
    key: string;
    component: string;
    alttext: string;
}

export interface Subscription {
    displayeventsource: boolean;
}